+++ cdcagg_client/sync.py
@@ -14,10 +14,7 @@
 """Synchronize records to CESSDA CDCAggregator Document Store"""
 import sys
 import logging
-from kuha_common.query import (
-    QueryController,
-    FilterKeyConstants
-)
+from kuha_common.query import QueryController
 from kuha_common import (
     conf,
     cli_setup
@@ -48,15 +45,30 @@ class StudyMethods(kuha_client.CollectionMethods):
     async def query_record(self, record):
         """Query record from Document Store.
 
+        This method is called from upsert() method. If this method
+        returns None, then the upsert() will never call
+        update_record(), but will call create_record() instead.
+
         :param :obj:`Study` record: Study to query for.
         :returns: Result of the query.
         :rtype: Instance of Study or None.
         """
-        return await QueryController().query_single(Study, _filter={
-            FilterKeyConstants.and_: [
-                {Study._provenance.attr_direct: True},
-                {Study._provenance.attr_identifier: record._provenance[-1].attr_identifier.get_value()},
-                {Study._provenance.attr_base_url: record._provenance[-1].attr_base_url.get_value()}]})
+        # This query uses elemMatch to make sure the identifier and base_url are
+        # within the same provenance item. Direct attribute here is not needed
+        # since we're interested in indirect records as well.
+
+        # Loop throught all provenance items to query for a record with matching
+        # provenance base_url + identifier combination.
+        rec = None
+        for prov in record._provenance:
+            rec = await QueryController().query_single(Study, _filter={
+                Study._provenance: {
+                    QueryController.fk_constants.elem_match: {
+                        Study._provenance.attr_base_url: prov.attr_base_url.get_value(),
+                        Study._provenance.attr_identifier: prov.attr_identifier.get_value()}}})
+            if rec:
+                break
+        return rec
 
     async def query_distinct_ids(self):
         """Query distinct IDs from collection that are not deleted.